+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## [Unreleased]
 
+- overwriteRequestHeaders - Allow zero value
+- overwriteRequestPathVariables - Allow zero value
+- overwriteRequestQueryParams - Allow zero value (#603)
 - Fuzzing - Improve fuzzing with zero values (#599)
 - Improved OpenAPI parsing error reporting (#591)
 - Replaced faker with @faker-js (#593)